extracted SourcesFromFS and constants
